Will/16 create enemies (#9)
Added Basic Enemy AI to move to player + wall collisions.

Added smarter enemy AI that tracks the player and
goes around walls. Shoots at player.

The enemies stops and shoots at the player for certain
durations. The rate of fire is 2-3.

The enemies has states for Roaming and Pursing.

fix: Fixed bug when enemies get stuck on the corner
of a wall. Using context steering algorithm.

---------

Co-authored-by: CamoDuck <maysamf@students.cs.ubc.ca>